Fixup reinstall/install logic
Failure of a single {spl,zfs} RPM to reinstall would trigger the
install logic which would NOT replace kmod-{spl,zfs} RPM if it was
installed. This changes the behavior to attempt to reinstall each RPM
and if that fails (because it's not installed) to install it. Also,
make the build/install less chatty as all output ends up in
/var/log/messages.